backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httptools==0.6.1
idna==3.6
pydantic==2.6.3
pydantic_core==2.16.3
pyspark==3.3.1
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
python-multipart==0.0.9
MarkupSafe==2.1.5
filelock==3.13.1
fsspec==2024.3.0
jinja2==3.1.3
mpmath==1.3.0
networkx==3.2.1
sympy==1.12
torch==2.2.1
SQLAlchemy==2.0.28
aiohttp==3.9.3 
aiosignal==1.3.1 
async==timeout==4.0.3 
attrs==23.2.0 
certifi==2024.2.2 
charset==normalizer==3.3.2 
dataclasses==json==0.6.4 
frozenlist==1.4.1 
greenlet==3.0.3 
jsonpatch==1.33 
jsonpointer==2.4 
langchain==0.1.12 
langchain-community==0.0.28 
PyPDF2==3.0.1
langchain-core==0.1.32 
langchain-text-splitters==0.0.1 
langsmith==0.1.27 
marshmallow==3.21.1 
multidict==6.0.5 
mypy-extensions==1.0.0 
numpy==1.26.4 
orjson==3.9.15 
packaging==23.2 
requests==2.31.0 
tenacity==8.2.3 
typing-inspect==0.9.0 
urllib3==2.2.1 
yarl==1.9.4
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==28.2
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
langchain-google-genai==0.0.11
huggingface-hub==0.21.4 
regex==2023.12.25 
safetensors==0.4.2 
tokenizers==0.15.2 
transformers==4.38.2
Pillow==10.2.0 
joblib==1.3.2 
scikit-learn==1.4.1.post1 
scipy==1.12.0 
sentence-transformers==2.5.1 
threadpoolctl==3.3.0
asgiref==3.7.2 
backoff==2.2.1 
bcrypt==4.1.2 
build==1.1.1 

chroma-hnswlib==0.7.3 
chromadb==0.4.24 
coloredlogs==15.0.1 
deprecated==1.2.14 
flatbuffers==24.3.7 
humanfriendly-10.0 
importlib-metadata==6.11.0 
importlib-resources==6.3.1 
kubernetes==29.0.0 
mmh3==4.1.0 
monotonic==1.6 
oauthlib==3.2.2 
onnxruntime==1.17.1 
opentelemetry-api==1.23.0 
opentelemetry-exporter-otlp-proto-common==1.23.0 
opentelemetry-exporter-otlp-proto-grpc==1.23.0 
opentelemetry-instrumentation==0.44b0 
opentelemetry-instrumentation-asgi==0.44b0 
opentelemetry-instrumentation-fastapi==0.44b0 
opentelemetry-proto==1.23.0 
opentelemetry-sdk==1.23.0 
opentelemetry-semantic-conventions==0.44b0 
opentelemetry-util-http==0.44b0 
overrides==7.7.0 
posthog==3.5.0 
pulsar-client==3.4.0 
pypika==0.48.9 
pyproject_hooks==1.0.0 
pyreadline3==3.4.1 
python-dateutil==2.9.0.post0 
requests-oauthlib==1.4.0 
six==1.16.0 
tomli==2.0.1 
typer==0.9.0 
websocket-client==1.7.0 
wrapt==1.16.0 
zipp==3.18.1
faiss-cpu==1.8.0
pypdf==4.1.0
docx2txt==0.8
lxml==5.1.0 
python==docx-1.1.0